+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 - fix invalid sampler warning
 - fix starting from non git repo
 - fix control api negative prompt handling
+- fix saving style without name provided
 - restructure api examples: `cli/api-*`
 - handle theme fallback when invalid theme is specified
 - remove obsolete training code leftovers